supplement and optimization

+++ include/SysConfig.h
@@ -35,7 +35,7 @@ class SysConfig {
     int16_t          lora_fcnt       = 0;
     int16_t          lora_sucess_cnt = 0;
     uint8_t          sd_status       = 0; // 0- not init, 1-  connected, 2- sd full
-    uint8_t          sensor_save_flag = 0;
+    uint16_t         sensor_save_flag = 0;
 
     static SemaphoreHandle_t lock;
